+++ Source/DynamicScene/CzmlDefaults.js
@@ -12,6 +12,7 @@ define([
         './DynamicPolygon',
         './DynamicPolyline',
         './DynamicPyramid',
+        './DynamicScreenOverlay',
         './DynamicBillboardVisualizer',
         './DynamicEllipsoidVisualizer',
         './DynamicConeVisualizerUsingCustomSensor', //CZML_TODO Replace with './DynamicConeVisualizer', once ComplexConicSensor works.
@@ -20,7 +21,8 @@ define([
         './DynamicPointVisualizer',
         './DynamicPolygonVisualizer',
         './DynamicPolylineVisualizer',
-        './DynamicPyramidVisualizer'
+        './DynamicPyramidVisualizer',
+        './DynamicScreenOverlayVisualizer'
     ], function(
         DynamicObject,
         DynamicBillboard,
@@ -34,6 +36,7 @@ define([
         DynamicPolygon,
         DynamicPolyline,
         DynamicPyramid,
+        DynamicScreenOverlay,
         DynamicBillboardVisualizer,
         DynamicEllipsoidVisualizer,
         DynamicConeVisualizer,
@@ -42,7 +45,8 @@ define([
         DynamicPointVisualizer,
         DynamicPolygonVisualizer,
         DynamicPolylineVisualizer,
-        DynamicPyramidVisualizer) {
+        DynamicPyramidVisualizer,
+        DynamicScreenOverlayVisualizer) {
     "use strict";
 
     /**
@@ -76,6 +80,7 @@ define([
                     DynamicPolygon.processCzmlPacket,
                     DynamicPolyline.processCzmlPacket,
                     DynamicPyramid.processCzmlPacket,
+                    DynamicScreenOverlay.processCzmlPacket,
                     DynamicObject.processCzmlPacketPosition,
                     DynamicObject.processCzmlPacketViewFrom,
                     DynamicObject.processCzmlPacketOrientation,
@@ -99,6 +104,7 @@ define([
                    DynamicPolygon.mergeProperties,
                    DynamicPolyline.mergeProperties,
                    DynamicPyramid.mergeProperties,
+                   DynamicScreenOverlay.mergeProperties,
                    DynamicObject.mergeProperties],
 
         /**
@@ -117,6 +123,7 @@ define([
                     DynamicPolygon.undefineProperties,
                     DynamicPolyline.undefineProperties,
                     DynamicPyramid.undefineProperties,
+                    DynamicScreenOverlay.undefineProperties,
                     DynamicObject.undefineProperties,
                     DynamicClock.undefineProperties],
 
@@ -137,7 +144,8 @@ define([
                     new DynamicPolygonVisualizer(scene),
                     new DynamicPolylineVisualizer(scene),
                     new DynamicPyramidVisualizer(scene),
-                    new DynamicPathVisualizer(scene)];
+                    new DynamicPathVisualizer(scene),
+                    new DynamicScreenOverlayVisualizer(scene)];
         }
     };